+++ loan-service/src/main/java/dev/kilima/training/loan/service/LoanService.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import dev.kilima.training.loan.entity.LoanDetails;
 
 public interface LoanService {
-	Optional<Integer> getMaxId();
+	Optional<Long> getMaxId();
 	String checkLoanStatus(long loanId);
-	int applyForLoan(LoanDetails loan);
+	long applyForLoan(LoanDetails loan);
 	
 }
